misc: build: Adjust version.tcl and create build-info
create_build_info.sh will examine the main de1app repo working tree
(assumed to be the current working directory) for annotated tags and
determine an appropriate version number and version string based on
the most-recent tag matching "v*"

  tag:		     string:			Tcl:

  v1.36	     	       v1.36			1.36
  v1.36 + 3 commits    v1.36-3-gf6844170	1.36.0.3
  v1.36.1   	       v1.36.1			1.36.1
  v1.36.1 + 5 commits  v1.36.1-5-gac678df9	1.36.1.5

version.tcl is used as a template and is updated in-place
(no commit is made).

Various parameters, such as the GIT_ROOT and BUILD_TIMESTAMP, may be
overriden through environment variables. At this time, the script does
not generate or read timestamp.txt

If used in a development situation, additional suffixes may be added
that indicate if the tree is in a "dirty" state (-dirty), and/or has
files that are not known to git (+).

In addition, a de1app/build-info is generated containing infiormation
about the state of de1app repo and all the registered submodules.

build-info includes the equivalent version_string (though requirements
for annotated tags have been relaxed for submodules), the last commit
hash, and the last commit date (when committed to the repo, not the
"author date").

This information is tab-separated key-value pairs on each line. It is
suitable for machine extraction, as well as being readable directly
or in run-time logs.

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>